Fix capturing in HQ mode
This fix, has a temporary work around that only allows to capture 8MP images.

In the constructor of cameraHal we initialize the IPP with determined width and height, but when HQ mode is configured the MMS library modifies the
width and height of the image to be capture so it was causing a dsp mmu fault in the IPP algorithm.
The workaround is to initialize the modified width and height that the MMS library produces.