OMAP4 CameraHAL: Fix for resolution switch after image capture
This patch resolves the issue with switching between resolutions after
an image capture. Basically, the image capture port was not being disabled
after stopImageCapture. so, when the preview resolution got changed and
the component was transitioned to LOADED state and back to IDLE, it hung
because the image capture port is not provided with buffers.
Disabling the image capture port inside stopImageCapture ensured that
there are no requirements for transitioning to EXECUTING state other
than the preview port being provided with buffers.

Change-Id: Id8e9c5c43bbf5c12ca612caad0769e3009e05fbb
Signed-off-by: Sundar Raman <sunds@ti.com>