OMAP4:AudioHAL:Modem:Put Audio Modem HAL under a pthread
These changes is to prepare a change in the Audio Modem Vendor
HAL interface.
Thanks to this change, the Audio Modem Vendor HAL interface become
a blocking-capable interface.
Therefore when a Modem Vendor i/f is blocked,
the Audio HAL in not blocked.

This is a requirement from modem vendor to have the possibility
for example:

- to wait for call connection before to start audio modem.
- or to wait for call assignment to report which samplerate
  will be used during the call.

Signed-off-by: Francois Mazard <f-mazard@ti.com>

Patch set 5: Simple cherry-pick on current tree.  No changes made.
Taking approvals and verification from patch set 4.
Signed-off-by: Jean Johnson <jean-johnson@ti.com>
Change-Id: I5c9cf0b68d254f7736432da7b660d67d3873434f