hardware/ti/omap3/ti_st/uim-rfkill: wake lock HACK
Holding wake locks when BT/FM/GPS is on and release wake locks when
BT/FM/GPS off.
This is temporary hack to not allow suspend when BT/FM/GPS is on.
There is some issue in UART driver when core off is enabled, this patch
will be reverted once UART driver issue resolved.

Change-Id: Iaad844b57befbe2e94037d3b98ee07d9d001e1e4
Signed-off-by: Bheemsen Kulkarni <x0099674@ti.com>
Signed-off-by: Pavan Savoy <pavan_savoy@ti.com>