OMAP4 CameraHal: Fix tearing issue with temporal bracketing in cases where setParameter is called
If setParameter is called when temporal bracketing is enabled or in general when Camera is in EXECUTING state,
it was resulting in tearing effect which is easily observed in temporal bracketing usecase. Both the snapshot
and the captured JPEG had tearing in it.

The reason was GBCE configuration on the fly (setting to OFF in HS mode) was causing this issue.
Also, With each setParameter, GBCE, face detection and GLBCE were configured even though the setting
hasn't been changed since last time.

This patch removes GBCE, GLBCE configuration for all other modes than HQ modes as these are valid only for HQ.
Also, adds a check to see if the configuration from the user has changed before reconfiguring OMXCamera
for the following settings - FD, GBCE, GLBCE.

Change-Id: I7573967844d7717c463834f8285f9adb186e1dcd
Signed-off-by: Sundar <sunds@ti.com>