libOverlay fix for setParameter and color key handling
Description: Since there is no way to disable the overlay parameters,
The control path setParameter is reused to disble the feature if the value
of a particular parameter is requested as -ve. In order to support this,
the setParameter is modified to accept -ve integers as the parameter values.
And also the commit call bug is fixed wrt color key handling.

Validation# tested AV playback from Gallery and Camera preview

DRT #
OMAPS00229539:PT Ad-hoc: Artifacts appears when camera app is been launched

Change-Id: I56c69715026609dd4e495de0560736fcb82c4cf4
Signed-off-by: Sunita Nadampalli <sunitan@ti.com>