OMAP4 CameraHAL: Configure picture resolution to be same as preview resolution during preview, if capture mode is not set
This patch sets the picture resolution to be the same as the capture resolution if the capture mode is not
explicitly set by the application.

This is needed to provide compatibility for 3rd party apps that perform video recording but inadvertently set the
picture resolution which leads to black bars in the default HQ mode.

Validated using a modified version of OMAP4Camera app by not explicitly configuring the capture mode from the app.
See no black bars to correct for the aspect ratio mismatch during video record.
Tested till 1080p.

Change-Id: I9b23bf5b745af403146cddcb88c1cec85bd0eb0d
Signed-off-by: Sundar <sunds@ti.com>