liboverlay changes for resetting the manager to primary lcd
Description# In order to brign the overlay back to default settings
while closing, the manager is getting reset to "lcd".
Earlier this reset was happening the control context destroy with the assumption
that control context destroy is the last call.
But, because of race condition between surface flinger and media server,
it is possible the control context destroy can happen even before the
data context. Hence the manager is reset in the middle and hence left at null.
Inorder to avoid this scenario, the manager reset is done in initialization
which is guaranteed to be sequential.

Validation# tested AV playback on HDMI and LCD with dynamic switching.

Signed-off-by: Sunita Nadampalli <sunitan@ti.com>

Patch Set 3: Cherry-pick on current tree head to break unnecessary
dependency.  No changes made. Taking verifications and approvals from
patch set 2.
Signed-off-by: Jean Johnson <jean-johnson@ti.com>
Change-Id: I190ed515259a500a3d519a771d04f61cbc9123c4