OMAP4 CameraHAL: Fix AF timeout issue
In night mode for exposure and scene, frame rate is very low and hence AF takes time to converge.
Since the AF timeout is only 3 seconds within Camera adapter, CameraHAL timesout. But the callback
from Ducati arrives later and tries to signal the semaphore which is no longer valid after exiting
doFocus menthod and this crashes Media Server.

This issue is fixed by the following two changes
1. Increased AF timeout to 10 seconds, so that the AF has enough time to converge in Night Mode.
The maximum latency seen so far in Night mode is 5.2 seconds.
2. Signal a dummy AF event in case of timeout and also cancel the AF callback request from Ducati.
The dummy AF event is needed just in case the callback comes even after disabling it because of
timing of the cancel call.

Signed-off-by: Sundar <sunds@ti.com>

Patch Set 2: Cherry-pick on top of current tree.  No changes made.  Taking
verifications and approvals from patch set 1.
Signed-off-by: Jean Johnson <jean-johnson@ti.com>

Change-Id: I95fc7536bb3f4777d7fff5d4cb4f6b3e640d31b3