OMAP4:TIHardwareRenderer changes to use codec recommended buffer count for overlay
Description: TIHardware Renderer is modified to receive the number of buffers
to be created by the overlay. The constructor is overloaded to receive
the extra parameter from the awesome player. This patch also involves
the removal of lower limit for the number of buffers. Hence, the number
of overlay buffers requested is purely dictated by the codec.

Validation: tested AV playback from Android Gallery app for files of
different resolutions.

Patchset3:
 - Author:Deva
 - Use number of output buffers as recommended by user/codec.
 - Validation of ducati/sw codec playabck

Change-Id: I76ab8a34fc0462554f778a3d832459c04c7343c4
Signed-off-by: Sunita Nadampalli <sunitan@ti.com>
Signed-off-by: Devaraj Rangasamy <dev@ti.com>