Thermal Manager: Make ThermalDaemon not dependent on hwmon instance
To make the thermal daemon more generic and not dependent on HWMON instance
number, this patch replaces usage of /sys/class/hwmon/hwmonx/device path by
direct access to device driver instance when it is possible.
When it is not possible, the detection of uevent source excludes the HWMON
instance number.

Change-Id: Ic9db1087fd73f7c86e6b2568e0ea4f881039a257
Signed-off-by: Sebastien Sabatier <s-sabatier1@ti.com>