Thermal Manager: Use PCB temp for extrapolation of hot spot temp
Even if OMAP4460/4470 on-die temperature sensor is located as close
as possible to OMAP MPU-SS, there is still an offset between on-die
sensor and OMAP hot spot.
This offset depends a lot on the additional system power:
various laws could be defined depending on the system power.
Relying on a PCB sensor allows correcting this offset temperature:
linear law can be used in thermal policy to adjust the temperature
read on the sensor and correct the offset at run time.
This requires to have PCB sensor close to OMAP die: this config
is available on TI OMAP4460 ES1.1 SOM (processor board).

This patch adds the support of PCB temperature:
  - Added parameters in libthermal-manager.cfg to support or not the
    external PCB sensor
    This allows keeping backward compatibility with current
    implementation
  - Added support of external Thermistor on OMAP4460 ES1.1 SOM
    This requires as well a dedicated kernel driver for TWL6030_GPADC
  - Updated formula to extrapolate hot spot from on-die sensor
    and PCB sensor
  - Updated reconfiguration of thresholds that should take into
    account the variation of PCB temperature
  - Added additional thread to monitor the on-die temperature and make
    average on latest 20 measurements:
    This average temperature is mandatory to take care of huge delta
    between on-die sensor and PCB sensor when the temperature
    increases quickly (burst CPU activity).
    This huge delta is coming from different response time between
    on-die sensor and PCB sensor.

Change-Id: I36dae8008d8ed9bfad4a6e6e54cf9a4aa297892a
Signed-off-by: Sebastien Sabatier <s-sabatier1@ti.com>