Thermal Manager: manage dynamically the support of PCB temp
This patch checks if from SW Kernel point-of-view the PCB temperature sensor
is accessible.
If it is not accessible and even if configuration file requests the support
of this PCB temperature sensor, then the thermal manager disables dynamically
the support of the PCB temperature sensor and force some parameters.
This patch does not protect against bad location of the PCB temperature sensor.

Change-Id: Ie797aab6cb81683a29531598b8011487607db435
Signed-off-by: Sebastien Sabatier <s-sabatier1@ti.com