Fixed Thermal Daemon behavior on 4430.
Attempt to open non existent item in sysfs, resulted in exit(), without visible messages.
This happened due to atempts to read sysfs files, that are not present on 4430.
So, I added error messages to read_from_file(), and moved call to read_from_file() to main() of ThermalDaemon.
So, monitoring thread is starting only if appropriate items exist in sysfs.

Change-Id: I348df989d7290f8fc65fb05071f1feddb5c7b93c
Signed-off-by: Ruslan Shymkevych <x0153342@ti.com>